Misc. bug fixes and refactorings, and the read-only mapped fs seems to be working. JRuby example working again, OSGi almost there

+++ src/edu/ubc/mirrors/BoxingFieldMirror.java
@@ -21,7 +21,7 @@ public abstract class BoxingFieldMirror implements FieldMirror {
 
     public short getShort() throws IllegalAccessException {
         Object value = getBoxedValue();
-        return value == null ? 0 : ((Byte)value).byteValue();
+        return value == null ? 0 : ((Short)value).byteValue();
     }
     
     public int getInt() throws IllegalAccessException {
